+++ NEWS.md
@@ -1,3 +1,14 @@
+# (upcoming)
+
+##### Support optional diagnostic tags ([#794][github#794])
+A [diagnostic tag][diagnostictag] can indicate either "unused or
+unnecessary code" or "deprecated or obsolete code".  Following the
+rendering suggestions in the protocol, we fade out unnecessary code
+and strike-through deprecated code.
+
+##### New servers have been added to `eglot-server-programs`
+- racket-langserver ([#694][github#694])
+
 # 1.8 (12/1/2022)
 
 ##### Multiple servers supported out-of-box for same major mode ([#688][github#688])
@@ -42,7 +53,7 @@ elisp.
 If non-nil, Eglot will not send the Emacs process id to the language server.
 This can be useful when using docker to run a language server.
 
-##### Several new servers have been added to `eglot-server-programs`.
+##### Several new servers have been added to `eglot-server-programs`
 - cmake-language-server ([#787][github#787])
 - css-languageserver ([#204][github#204], [#769][github#769])
 - fortls ([#603][github#603])
@@ -216,6 +227,7 @@ TCP connection finds a listening server.
 
 ##### Assorted bugfixes
 
+[diagnostictag]: https://microsoft.github.io/language-server-protocol/specifications/specification-3-17/#diagnosticTag
 <!--- Now a bunch of references that I auto-generate with
 
 (cl-loop
@@ -304,6 +316,7 @@ and now said bunch of references-->
 [github#646]: https://github.com/joaotavora/eglot/issues/646
 [github#686]: https://github.com/joaotavora/eglot/issues/686
 [github#688]: https://github.com/joaotavora/eglot/issues/688
+[github#694]: https://github.com/joaotavora/eglot/issues/694
 [github#695]: https://github.com/joaotavora/eglot/issues/695
 [github#721]: https://github.com/joaotavora/eglot/issues/721
 [github#722]: https://github.com/joaotavora/eglot/issues/722
@@ -313,3 +326,4 @@ and now said bunch of references-->
 [github#751]: https://github.com/joaotavora/eglot/issues/751
 [github#769]: https://github.com/joaotavora/eglot/issues/769
 [github#787]: https://github.com/joaotavora/eglot/issues/787
+[github#794]: https://github.com/joaotavora/eglot/issues/794